modularizacion y finalizacion del proyecto con todas las pruebas completadas
