+++ src/js/constants.js
@@ -1,0 +1,19 @@
+export const SELECTORS = {
+    MAIN: document.querySelector('.main'),
+    FOOTER: document.querySelector('.footer'),
+    NEW_TODO_INPUT: document.querySelector('.new-todo'),
+    TODO_LIST: document.querySelector('.todo-list'),
+    TODO_COUNT_SPAN: document.querySelector('.todo-count'),
+    CLEAR_COMPLETED_BUTTON: document.querySelector('.clear-completed'),
+
+};
+
+export const FILTERS = {
+    ALL_FILTER_LINK: document.querySelector('.filters a[href="#/"]'),
+    PENDING_FILTER_LINK: document.querySelector('.filters a[href="#/pending"]'),
+    COMPLETED_FILTER_LINK: document.querySelector('.filters a[href="#/completed"]'),
+}
+
+
+
+
