--- src/js/selectors.js
@@ -1,4 +1,0 @@
-export const main = document.querySelector('.main');
-export const footer = document.querySelector('.footer');
-export const newTodoInput = document.querySelector('.new-todo');
-export const firstLi = document.querySelector('ul.todo-list li')
